[SWM-227] Fix : uri /lectures/courses/{courseId} to courses/{courseId}
and change week to section
