+++ src/main/java/com/m9d/sroom/lecture/service/LectureServiceV2.java
@@ -367,8 +367,4 @@ public class LectureServiceV2 {
             throw new VideoIndexParamException();
         }
     }
-
-    public CourseDetail getCourseDetail(Long memberId, Long courseId) {
-        return new CourseDetail();
-    }
 }
